Added simulate() function
Added code for simulate() function including prints of results and saving of results. Difficulty of track is calculated as the length of the track multiplied by the number of turns divided by the length of the longest straight. It's random but its a number, if anyone has a better suggestion feel free to change it.