Started reworking Logistics
Introduced a Facade pattern for Team in order to facilitate communication between Logistics and Engineering. Also introduced EquipmentComposite as a Composite pattern. Factory methods were introduced in order to create the different types of Equipment, as well as to create the different modes of Transport. Design for Shipment, Container and LogisticsTeam is still incomplete